--- a/clients/include/testing_gbmv_batched.hpp
+++ b/clients/include/testing_gbmv_batched.hpp
@@ -168,9 +168,9 @@ void testing_gbmv_batched(const Arguments& arg)
     rocblas_local_handle handle;
 
     // argument sanity check before allocating invalid memory
-    bool invalidSize = M < 0 || N < 0 || lda < KL + KU + 1 || !incx || !incy || batch_count < 0
-                       || KL < 0 || KU < 0;
-    if(invalidSize || !M || !N || !batch_count)
+    bool invalid_size = M < 0 || N < 0 || lda < KL + KU + 1 || !incx || !incy || batch_count < 0
+                        || KL < 0 || KU < 0;
+    if(invalid_size || !M || !N || !batch_count)
     {
         EXPECT_ROCBLAS_STATUS(rocblas_gbmv_batched<T>(handle,
                                                       transA,
@@ -187,7 +187,7 @@ void testing_gbmv_batched(const Arguments& arg)
                                                       nullptr,
                                                       incy,
                                                       batch_count),
-                              invalidSize ? rocblas_status_invalid_size : rocblas_status_success);
+                              invalid_size ? rocblas_status_invalid_size : rocblas_status_success);
         return;
     }