--- a/clients/include/testing_ger_batched.hpp
+++ b/clients/include/testing_ger_batched.hpp
@@ -101,15 +101,14 @@ void testing_ger_batched(const Arguments& arg)
     rocblas_local_handle handle;
 
     // argument check before allocating invalid memory
-    if(M <= 0 || N <= 0 || lda < M || lda < 1 || !incx || !incy || batch_count <= 0)
+    bool invalidSize = M < 0 || N < 0 || lda < M || lda < 1 || !incx || !incy || batch_count < 0;
+    if(invalidSize || !M || !N || !batch_count)
     {
         EXPECT_ROCBLAS_STATUS(
             (rocblas_ger_batched<T, CONJ>(
-                handle, M, N, &h_alpha, nullptr, incx, nullptr, incy, nullptr, lda, batch_count)),
+                handle, M, N, nullptr, nullptr, incx, nullptr, incy, nullptr, lda, batch_count)),
 
-            M < 0 || N < 0 || lda < M || lda < 1 || !incx || !incy || batch_count < 0
-                ? rocblas_status_invalid_size
-                : rocblas_status_success);
+            invalidSize ? rocblas_status_invalid_size : rocblas_status_success);
         return;
     }