--- a/clients/include/testing_ger_strided_batched.hpp
+++ b/clients/include/testing_ger_strided_batched.hpp
@@ -129,8 +129,8 @@ void testing_ger_strided_batched(const Arguments& arg)
     size_t size_y   = N * abs_incy;
 
     // argument check before allocating invalid memory
-    bool invalidSize = M < 0 || N < 0 || lda < M || lda < 1 || !incx || !incy || batch_count < 0;
-    if(invalidSize || !M || !N || !batch_count)
+    bool invalid_size = M < 0 || N < 0 || lda < M || lda < 1 || !incx || !incy || batch_count < 0;
+    if(invalid_size || !M || !N || !batch_count)
     {
         EXPECT_ROCBLAS_STATUS((rocblas_ger_strided_batched<T, CONJ>(handle,
                                                                     M,
@@ -146,8 +146,8 @@ void testing_ger_strided_batched(const Arguments& arg)
                                                                     lda,
                                                                     stride_a,
                                                                     batch_count)),
-                              invalidSize || batch_count < 0 ? rocblas_status_invalid_size
-                                                             : rocblas_status_success);
+                              invalid_size || batch_count < 0 ? rocblas_status_invalid_size
+                                                              : rocblas_status_success);
 
         return;
     }