--- a/clients/include/testing_hbmv_batched.hpp
+++ b/clients/include/testing_hbmv_batched.hpp
@@ -157,22 +157,23 @@ void testing_hbmv_batched(const Arguments& arg)
     rocblas_local_handle handle;
 
     // argument sanity check before allocating invalid memory
-    if(N <= 0 || K < 0 || lda <= K || !incx || !incy || batch_count <= 0)
+    bool invalidSize = N < 0 || K < 0 || lda <= K || !incx || !incy || batch_count < 0;
+    if(invalidSize || !N || !batch_count)
     {
-        static const size_t    safe_size = 100; // arbitrarily set to 100
-        device_batch_vector<T> dA1(safe_size, 1, 5);
-        device_batch_vector<T> dx1(safe_size, 1, 5);
-        device_batch_vector<T> dy1(safe_size, 1, 5);
-        CHECK_HIP_ERROR(dA1.memcheck());
-        CHECK_HIP_ERROR(dx1.memcheck());
-        CHECK_HIP_ERROR(dy1.memcheck());
-
-        EXPECT_ROCBLAS_STATUS(
-            rocblas_hbmv_batched<T>(
-                handle, uplo, N, K, &h_alpha, dA1, lda, dx1, incx, &h_beta, dy1, incy, batch_count),
-            (N < 0 || K < 0 || lda <= K || !incx || !incy || batch_count < 0)
-                ? rocblas_status_invalid_size
-                : rocblas_status_success);
+        EXPECT_ROCBLAS_STATUS(rocblas_hbmv_batched<T>(handle,
+                                                      uplo,
+                                                      N,
+                                                      K,
+                                                      nullptr,
+                                                      nullptr,
+                                                      lda,
+                                                      nullptr,
+                                                      incx,
+                                                      nullptr,
+                                                      nullptr,
+                                                      incy,
+                                                      batch_count),
+                              invalidSize ? rocblas_status_invalid_size : rocblas_status_success);
 
         return;
     }