--- a/clients/include/testing_hbmv_strided_batched.hpp
+++ b/clients/include/testing_hbmv_strided_batched.hpp
@@ -191,27 +191,26 @@ void testing_hbmv_strided_batched(const Arguments& arg)
     rocblas_local_handle handle;
 
     // argument sanity check before allocating invalid memory
-    if(N <= 0 || K < 0 || lda <= K || !incx || !incy || batch_count <= 0)
+    bool invalidSize = N < 0 || K < 0 || lda <= K || !incx || !incy || batch_count < 0;
+    if(invalidSize || !N || !batch_count)
     {
         EXPECT_ROCBLAS_STATUS(rocblas_hbmv_strided_batched<T>(handle,
                                                               uplo,
                                                               N,
                                                               K,
-                                                              &h_alpha,
+                                                              nullptr,
                                                               nullptr,
                                                               lda,
                                                               stride_A,
                                                               nullptr,
                                                               incx,
                                                               stride_x,
-                                                              &h_beta,
+                                                              nullptr,
                                                               nullptr,
                                                               incy,
                                                               stride_y,
                                                               batch_count),
-                              (N < 0 || K < 0 || lda <= K || !incx || !incy || batch_count < 0)
-                                  ? rocblas_status_invalid_size
-                                  : rocblas_status_success);
+                              invalidSize ? rocblas_status_invalid_size : rocblas_status_success);
 
         return;
     }