--- a/clients/include/testing_hbmv_strided_batched.hpp
+++ b/clients/include/testing_hbmv_strided_batched.hpp
@@ -191,8 +191,8 @@ void testing_hbmv_strided_batched(const Arguments& arg)
     rocblas_local_handle handle;
 
     // argument sanity check before allocating invalid memory
-    bool invalidSize = N < 0 || K < 0 || lda <= K || !incx || !incy || batch_count < 0;
-    if(invalidSize || !N || !batch_count)
+    bool invalid_size = N < 0 || K < 0 || lda <= K || !incx || !incy || batch_count < 0;
+    if(invalid_size || !N || !batch_count)
     {
         EXPECT_ROCBLAS_STATUS(rocblas_hbmv_strided_batched<T>(handle,
                                                               uplo,
@@ -210,7 +210,7 @@ void testing_hbmv_strided_batched(const Arguments& arg)
                                                               incy,
                                                               stride_y,
                                                               batch_count),
-                              invalidSize ? rocblas_status_invalid_size : rocblas_status_success);
+                              invalid_size ? rocblas_status_invalid_size : rocblas_status_success);
 
         return;
     }