--- a/clients/include/testing_hemv_batched.hpp
+++ b/clients/include/testing_hemv_batched.hpp
@@ -149,31 +149,22 @@ void testing_hemv_batched(const Arguments& arg)
     rocblas_local_handle handle;
 
     // argument sanity check before allocating invalid memory
-    if(N < 0 || lda < N || lda < 1 || !incx || !incy || batch_count <= 0)
+    bool invalidSize = N < 0 || lda < N || lda < 1 || !incx || !incy || batch_count < 0;
+    if(invalidSize || !N || !batch_count)
     {
-        static const size_t    safe_size = 100; // arbitrarily set to 100
-        device_batch_vector<T> dA1(safe_size, 1, 5);
-        device_batch_vector<T> dx1(safe_size, 1, 5);
-        device_batch_vector<T> dy1(safe_size, 1, 5);
-        CHECK_HIP_ERROR(dA1.memcheck());
-        CHECK_HIP_ERROR(dx1.memcheck());
-        CHECK_HIP_ERROR(dy1.memcheck());
-
         EXPECT_ROCBLAS_STATUS(rocblas_hemv_batched<T>(handle,
                                                       uplo,
                                                       N,
-                                                      &h_alpha,
-                                                      dA1.ptr_on_device(),
+                                                      nullptr,
+                                                      nullptr,
                                                       lda,
-                                                      dx1.ptr_on_device(),
+                                                      nullptr,
                                                       incx,
-                                                      &h_beta,
-                                                      dy1.ptr_on_device(),
+                                                      nullptr,
+                                                      nullptr,
                                                       incy,
                                                       batch_count),
-                              (N < 0 || lda < N || lda < 1 || !incx || !incy || batch_count < 0)
-                                  ? rocblas_status_invalid_size
-                                  : rocblas_status_success);
+                              invalidSize ? rocblas_status_invalid_size : rocblas_status_success);
 
         return;
     }