--- a/clients/include/testing_her_batched.hpp
+++ b/clients/include/testing_her_batched.hpp
@@ -66,13 +66,13 @@ void testing_her_batched(const Arguments& arg)
     rocblas_local_handle handle;
 
     // argument check before allocating invalid memory
-    if(N <= 0 || !incx || lda < 1 || lda < N || batch_count <= 0)
+    bool invalidSize = N < 0 || lda < 1 || lda < N || !incx || batch_count < 0;
+    if(invalidSize || !N || !batch_count)
     {
         EXPECT_ROCBLAS_STATUS(
             rocblas_her_batched<T>(
                 handle, uplo, N, nullptr, nullptr, incx, nullptr, lda, batch_count),
-            N < 0 || !incx || lda < 1 || lda < N || batch_count < 0 ? rocblas_status_invalid_size
-                                                                    : rocblas_status_success);
+            invalidSize ? rocblas_status_invalid_size : rocblas_status_success);
         return;
     }