… clients/gtest/gbmv_gtest.yaml
… clients/gtest/gemv_gtest.yaml
… clients/gtest/ger_gtest.yaml
… clients/gtest/geruc_gtest.yaml
… clients/gtest/hemv_gtest.yaml
… clients/gtest/her2_gtest.yaml
… clients/gtest/her_gtest.yaml
… clients/gtest/hpmv_gtest.yaml
… clients/gtest/hpr2_gtest.yaml
… clients/gtest/hpr_gtest.yaml
… clients/gtest/sbmv_gtest.yaml
… clients/gtest/spmv_gtest.yaml
… clients/gtest/spr2_gtest.yaml
… clients/gtest/spr_gtest.yaml
… clients/gtest/symv_gtest.yaml
… clients/gtest/syr2_gtest.yaml
… clients/gtest/syr_gtest.yaml
… clients/gtest/tbmv_gtest.yaml
… clients/gtest/tbsv_gtest.yaml
… clients/gtest/tpmv_gtest.yaml
… clients/gtest/tpsv_gtest.yaml
… clients/gtest/trmv_gtest.yaml
… clients/gtest/trsv_gtest.yaml
… clients/include/testing_hemv_strided_batched.hpp
… clients/include/testing_her2_batched.hpp
… clients/include/testing_her2_strided_batched.hpp
… clients/include/testing_her2k.hpp
… clients/include/testing_her2k_batched.hpp
… clients/include/testing_her2k_strided_batched.hpp
… clients/include/testing_her_strided_batched.hpp