Reducing the number of repeated L2 tests.
(Sorry for the massive amount of changed files for this PR!)

- Moved duplicated invalid_arg and quick-return tests to their own section in *.yaml files to reduce repeated tests. These are done in the "quick" tests
- Reduced the incx/incy and alpha/beta combinations for some pre-checkin/nightly tests
- Other small changes

See [LWPMLSE-39](http://ontrack-internal.amd.com/browse/LWPMLSE-39) and my comment for a more detailed description and for some discussion.

After these changes, here are the testing times for the L2 functions as done on my local machine. In general, the timing is within or (often times) underneath the guidelines set out in the documentation by Andrew.

Note that sometimes the pre-checkin tests take longer than the nightly tests. See the ticket for the reason for this. If this isn't acceptable we should have a discussion on what each set of tests (quick/pre-checkin/nightly) should _mean_ and what they should each be testing.

Function | Quick (s)  | Pre-Checkin (s) | Night (s) | Total (s)
-- | -- | -- | -- | --
GEMV | 2 | 30 | 60 | 92
GER | 12 | 83 | 438 | 533
HBMV | 8 | 17 | 16 | 41
HEMV | 15 | 20 | 16 | 51
HER | 3 | 36 | 50 | 89
HER2 | 7 | 14 | 63 | 84
HER2K | 1 | 4 | 44 | 49
HERK | 1 | 3 | 69 | 73
HERKX | 0 | 4 | 133 | 137
HPMV | 13 | 12 | 9 | 34
HPR | 6 | 68 | 34 | 108
HPR2 | 6 | 33 | 344 | 383
SBMV | 0 | 46 | 36 | 82
SPMV | 1 | 75 | 27 | 103
SPR | 14 | 58 | 42 | 114
SPR2 | 5 | 40 | 28 | 73
SYMV | 1 | 132 | 300 | 433
SYR | 8 | 33 | 201 | 242
SYR2 | 17 | 49 | 184 | 250
SYR2K | 1 | 14 | 158 | 173
SYRK | 1 | 2 | 324 | 327
SYRKX | 1 | 4 | 125 | 130
TBMV | 9 | 51 | 59 | 119
TBSV  | 22  | 21  | 228  | 271
TPMV | 2 | 11 | 129 | 142
TPSV | 27  | 7  | 204  | 238
TRMV | 3 | 20 | 236 | 259
TRSV | 17  | 42  | 480  | 539
**All L2** | **203** | **929** | **4037** | **5169**

Where the targets for L2 functions are:

x | Quick | Pre-Checkin | Nightly
-- | -- | -- | --
L2 | 6-36s | 35-100s | 200-650s