@daineAMD 2020-03-25
In which case is `lda < M || lda < 1` not equivalent to `lda < max(M, 1)`? When `M == 0` we have: `lda < max(M, 1) == lda < max(0, 1) == lda < 1` which means `lda == 0` would be an error in LAPACK as well. I think my logic is right on this one but I could be mistaken :)