@TorreZuk 2020-03-25
Sorry it seems I was muddled in my thinking I thought it was mismatched but yes the quick returns all happen afterwards so your code should match LAPACK for the values of interest.